test(agents): the id checks join the parity gate that owns contract-to-surface agreement
They were a file of their own, which implied a separate concern. They are not
one: agenttoolparity_test.go already asserts that the contract and the registered
surface agree about WHICH VERBS EXIST, in both directions, and its header states
the reason — "the contract does not describe the agent tool surface, it IS the
agent tool surface". Agreeing about the argument a caller must name is the same
relationship one level down, so it belongs in the same file and under the same
argument.

The explanation of why the drift gate cannot reach this now sits beside the gate
it qualifies, which is where a reader of that gate needs it: gen-agentpolicy
lifts seven fields out of crm.yaml and argument names are not among them, and
structurally cannot be — on REST the id is a path parameter and the bodies carry
only reason, trigger and evidence — so every InputSchema is a hand-written
literal and `make drift` is vacuous about it by construction.

Also from the gate run, two conventions this branch had to learn:

notTheRecordsOwnID is declared a `gatekit:fixture` rather than wrapped in
gatekit.Waive, and the distinction is gatekit's own: a waiver is a ratified COST
asked about an offender, so it decays when the offence goes, while this map is
asked BEFORE the subject is known. gatekit.Waived's doc refuses exactly that
pre-filter shape because it discards the findings behind it. The fixture pin moves
in this change, as a classification made once is supposed to.

And "Named once" in a test doc read as a uniqueness claim to the claims census.
The sentence says what it actually asserts now: the spec is resolved rather than
indexed and checked to still carry a threshold, so a catalog rename fails there
instead of handing every case a report with no thresholds and staying green.

Both checks re-verified after the move rather than assumed: an unfamiliar
`case_id` still reds the sibling census, and qualify_lead's original `record_id`
with no `record_type` still reds the schema-only one.
