+++ backend/gates/gatecensus_test.go
@@ -52,7 +52,7 @@ const (
 	// Permitting the marker without counting it would let it spread quietly and
 	// reopen the class these rules close; pinned, every new one moves a number a
 	// reviewer sees.
-	wantFixtureAnnotations = 47
+	wantFixtureAnnotations = 48
 )
 
 // censusDecl is one package-level declaration this census governs — a map from
